pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel", "build"]
build-backend = "setuptools.build_meta"

[project]
name = "alchemyst-langchain"
version = "0.1.0"
description = "Alchemyst's LangChain-style Python utilities"
readme = "README.md"
license = { text = "MIT" }
authors = [{ name = "Alchemyst", email = "saloni@getalchemyst.com" }]
requires-python = ">=3.10"
dependencies = [
  "requests>=2.31",
  "langchain-core>=0.1.0"
]
classifiers = [
  "Programming Language :: Python :: 3",
  "License :: OSI Approved :: MIT License",
  "Operating System :: OS Independent",
]

[project.optional-dependencies]
dev = [
  "pytest>=7.0",
  "responses>=0.23",
  "build>=0.10",
  "twine>=4.0"
]